Extract IOpenInIDERequestHandler.GetStatusAsync into IStatusRequestHandler
