+++ src/IssueViz.Security.UnitTests/OpenInIDE/Http/StatusOwinRequestHandlerTests.cs
@@ -50,7 +50,7 @@ namespace SonarLint.VisualStudio.IssueVisualization.Security.UnitTests.OpenInIDE
         [TestMethod]
         public void ApiPath_ReturnsExpectedPath()
         {
-            var testSubject = new StatusOwinRequestHandler(Mock.Of<IOpenInIDERequestHandler>());
+            var testSubject = new StatusOwinRequestHandler(Mock.Of<IStatusRequestHandler>());
 
             testSubject.ApiPath.Should().Be("/status");
         }
@@ -63,7 +63,7 @@ namespace SonarLint.VisualStudio.IssueVisualization.Security.UnitTests.OpenInIDE
 
             var apiResponse = new StatusResponse(expectedIdeName, expectedDescription);
 
-            var apiHandlerMock = new Mock<IOpenInIDERequestHandler>();
+            var apiHandlerMock = new Mock<IStatusRequestHandler>();
             apiHandlerMock.Setup(x => x.GetStatusAsync()).Returns(Task.FromResult<IStatusResponse>(apiResponse));
 
             var responseStream = new MemoryStream();
